% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_gaussian}
\alias{cov_gaussian}
\title{Covariance matrix of GP with Gaussian Kernel}
\usage{
cov_gaussian(X1, X2, tau = 1, s = 1)
}
\arguments{
\item{X1}{arma::mat, each row representing the coordiantes of a location}

\item{X2}{arma::mat, each row representing the coordiantes of a location}

\item{tau}{double, the time rescaling parameter of the GP}

\item{s}{double, the space rescaling parameter of the GP. Currently set as 1}
}
\value{
the computed covariance matrix.
}
\description{
This function evaluates the covariance matrix of GP with Gaussian Kernel using the coordinates.
}
