% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DAG_Chol}
\alias{DAG_Chol}
\title{Cholesky of DAG GP covariance matrix}
\usage{
DAG_Chol(X, dag, nu = 3/2, tau = 1, s = 1, cov_type = "matern", tol = 2e-16)
}
\arguments{
\item{X}{arma::mat, with each row representing the coordiante of a location}

\item{dag}{arma::field\url{arma::uvec}, each entry contains the indices of its parent sets in DAG ordering}

\item{nu}{double, the smoothness of the Matern process}

\item{tau}{double, rescaling parameter of the Matern process}

\item{tol}{double, in case of numerical singularity (negative conditional variance), set the conditional variance as tol}
}
\value{
Rcpp list consists of two entries, L and D, such that L D^{-1} L^T is the precision matrix.
D is the conditonal variance for ith element, while the off-diagonal elements of L\link{,i} are negative conditional regression coefficients of ith elements on its parents.
}
\description{
This function performs Cholesky decomposition of DAG GP covariance function. No input of the whole covariance matrix is
needed. Instead, this function will call evaluate the covariance matrix based on the DAG structure.
}
