% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGgrid.R
\name{DAGgrid}
\alias{DAGgrid}
\title{Norming DAG on grid}
\usage{
DAGgrid(Y, minsep, alpha = 3/2)
}
\arguments{
\item{Y}{A numeric vector or matrix, the observed response variable on the grid.}

\item{minsep}{A numeric value, Minimal separation distance of the grid.}

\item{alpha}{A numeric value, the smoothness of the Matern process, determining the cardinality of parent sets.}
}
\value{
A list of four entries:
X_ord: coordinates of the grid in DAG ordering;
dag_ord: a list of numeric vectors, each entry contains the indices of its parent sets in DAG ordering;
Y_ord: a numeric vector, the Y values in DAG ordering
sort_in_ord: a numeric vector, the ordering of DAG with respect to the coordinate ordering.
}
\description{
This function computes DAG for grid data. You do not need to provide coordinates
of grid, only the observed response variable on the grid. The size of the grid is
automatically extracted from the shape of the response Y.
}
