% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGgrid.R
\name{DAGgrid_per_1d}
\alias{DAGgrid_per_1d}
\title{Norming DAG on grid, 1-d}
\usage{
DAGgrid_per_1d(J, alpha)
}
\arguments{
\item{J}{An integer, number of layers. J determins the cardinality of the grid.}

\item{alpha}{A numeric value, the smoothness of the Matern process, determining the cardinality of parent sets.}

\item{sort_in_ord}{A numeric vector representing the ordering of elements in Xcoords. One can simply input the "sort_in_ord" variable from the output of the function DAGgrid_per_hd}
}
\value{
A list of three entries:
X_ord: coordinates of the grid in DAG ordering;
dag_ord: a list of numeric vectors, each entry contains the indices of its parent sets in DAG ordering;
sort_in_ord: a numeric vector, the ordering of DAG with respect to the coordinate ordering.
}
\description{
This function constructs norming DAG on a unidimesional grid of cardinality 2^J+1. For grid of other cardinality, please call
the function DAGgrid
}
