% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGgrid.R
\name{DAGgrid_test_1d}
\alias{DAGgrid_test_1d}
\title{Norming DAG for test data, 1-d grid algorithm}
\usage{
DAGgrid_test_1d(X, Xtest, alpha)
}
\arguments{
\item{Xtest}{A numeric matrix, each row corresponds to the coordinates of a test location}

\item{alpha}{A numeric value, the smoothness of the Matern process, determining the cardinality of parent sets.}

\item{Xcoords}{List of numeric vectors, each element of the list contains the coordinate values of training data}
}
\value{
A list of numeric vectors representing the test DAG. Each element of the list contains the indices of parents in DAG ordering.
}
\description{
This function constructs norming DAG on a test set using 1-d grid algorithm. Only
the training set is required to be on a grid, while test set can be arbitrary locations.
}
