% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{GPgrid}
\alias{GPgrid}
\title{DAG GP on grid}
\usage{
GPgrid(
  Y,
  minsep,
  adapt = FALSE,
  nu = 3/2,
  sig = 0.1,
  tau = 10,
  tau_bound = NULL,
  nug_prior = NULL,
  sig_bound = NULL,
  n_mcmc = 2000,
  n_burn = 1000,
  tol = 1e-12,
  ncgiter = 200,
  log_tau_fun = NULL
)
}
\arguments{
\item{Y}{numeric vector or numeric matrix, the response, with each element representing the value of response on its corresponding grid location.}

\item{minsep}{numeric value, the minimal separation distance of the grid.}

\item{adapt}{bool, whether adaptation on rescaling parameter is required. Default is FALSE.}

\item{nu}{numeric value, the smoothness of the Matern process}

\item{sig}{numeric value, initial value for the standard deviation of nugget effects}

\item{tau}{numeric value, initial value for rescaling parameter of the Matern process}

\item{tau_bound}{numeric vector of length 2, lower and upper bounds for tau}

\item{nug_prior}{numeric vector of length 2, parameters for inverse gamma prior for sigma^2}

\item{sig_bound}{numeric vector of length 2, lower and upper bounds for sigma, currently not used in the codes}

\item{n_mcmc}{unsigned integer, number of MCMC iterations}

\item{n_burn}{unsigned integer, number of MCMC iterations that are burnt}

\item{tol}{numeric value, the tolerance of squared l2 norm of conjugate gradient}

\item{ncgiter}{integer, the maximal number of conjugate gradient iterations}

\item{log_tau_prior_fun}{Function, log prior function for tau. Does not need to be specified if no adaptation is required.}
}
\value{
Rcpp list consists results of MCMC:
tau_mcmc, the MCMC samples of tau;
sig_mcmc, the MCMC samples of sig;
sig2_mean_mcmc, the mean value of sigma squared in each MCMC iteration;
Z_mcmc, a matrix, each row contains the MCMC sample of the latent function in that iteration;
Accept_ratio, average acceptance ratio for sampling the rescaling paramter tau. Is irrelevant if not in a hierarchical Bayes framework;
L & D, the ouput of function DAG_Chol in the last MCMC iterations;
RunTime: the total real world running time of the MCMC;
CG_iters: the number of conjuate gradient iterations for each MCMC step;
CG_err: the conjugate gradient error for each MCMC step;
X: coordinates of the grid, ordered by their coordinate values;
dag_ord: a list of numeric vectors, each entry contains the indices of its parent sets in DAG ordering;
Y: the input response Y;
sort_in_ord: a numeric vector, the ordering of DAG with respect to the coordinate ordering.
}
\description{
This is the high-level function to perform DAG GP on grid, including both building the DAG and the MCMC on the DAG.
}
