% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGgrid.R
\name{Rdag_to_Cppdag}
\alias{Rdag_to_Cppdag}
\title{Convert a R DAG to Rcpp DAG}
\usage{
Rdag_to_Cppdag(dag_ord)
}
\arguments{
\item{dag_ord}{A list, where each element of the list is a numeric vector containing indices of parents}
}
\value{
A list similar to input, except that all indices of parents get minus 1.
}
\description{
The indices in R start from 1, while indices in c++ start from 0. This function
simply minus 1 for all the parent sets indices in a dag.
}
